+++ src/types/eslint.d.ts
@@ -1,0 +1,5 @@
+import { ESLint } from "eslint";
+
+declare module "eslint" {
+  function loadESLint(): Promise<typeof ESLint>;
+}